@@ -11,17 +11,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.1.10",
-    "@angular/common": "^6.1.10",
-    "@angular/compiler": "^6.1.10",
-    "@angular/core": "^6.1.10",
-    "@angular/forms": "^6.1.10",
-    "@angular/http": "^6.1.10",
-    "@angular/platform-browser": "^6.1.10",
-    "@angular/platform-browser-dynamic": "^6.1.10",
-    "@angular/router": "^6.1.10",
+    "@angular/animations": "^11.1.2",
+    "@angular/common": "^11.1.2",
+    "@angular/compiler": "^11.1.2",
+    "@angular/core": "^11.1.2",
+    "@angular/forms": "^11.1.2",
+    "@angular/platform-browser": "^11.1.2",
+    "@angular/platform-browser-dynamic": "^11.1.2",
+    "@angular/router": "^11.1.2",
     "@jaortiz117/icon-map": "^2.2.0",
-    "bootstrap": "^4.3.1",
+    "@ng-bootstrap/ng-bootstrap": "^9.0.2",
+    "bootstrap": "^4.6.0",
     "chart.js": "^2.8.0",
     "core-js": "^2.6.9",
     "heatmap.js": "^2.0.5",
@@ -32,31 +32,32 @@
     "ngx-bootstrap": "^2.0.5",
     "pg": "^7.12.1",
     "pg-hstore": "^2.3.3",
-    "rxjs": "6.4.0",
+    "rxjs": "6.6.3",
     "sequelize": "^5.18.4",
+    "tslib": "^2.0.0",
     "ws": "^6.2.1",
-    "zone.js": "^0.8.29"
+    "zone.js": "~0.11.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.1101.2",
-    "@angular/cli": "^6.0.8",
-    "@angular/compiler-cli": "^6.1.10",
-    "@angular/language-service": "^6.1.10",
-    "@types/jasmine": "^2.8.16",
+    "@angular-devkit/build-angular": "~0.1101.2",
+    "@angular/cli": "^11.1.2",
+    "@angular/compiler-cli": "^11.1.2",
+    "@angular/language-service": "^11.1.2",
+    "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "^2.0.6",
-    "@types/node": "~8.9.4",
-    "codelyzer": "~4.2.1",
-    "jasmine-core": "~2.99.1",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~1.7.1",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^1.4.3",
-    "karma-jasmine": "~1.1.1",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^6.0.0",
+    "jasmine-core": "~3.6.0",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "~5.0.0",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~3.0.2",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
     "nodemon": "^1.19.2",
-    "protractor": "~5.3.0",
+    "protractor": "~7.0.0",
     "ts-node": "~5.0.1",
-    "tslint": "~5.9.1",
-    "typescript": "^2.7.2"
+    "tslint": "~6.1.0",
+    "typescript": "^4.1.3"
   }
 }
