@@ -9,7 +9,7 @@ const routes: Routes = [
 
 @NgModule({
   exports: [ RouterModule ],
-  imports: [ RouterModule.forRoot(routes) ]
+  imports: [ RouterModule.forRoot(routes, { relativeLinkResolution: 'legacy' }) ]
 })
 
 export class AppRoutingModule {
