@@ -17,6 +17,7 @@ import { SignalComponent } from './components/status-bar/signal/signal.component
 import { BuggysComponent } from './components/dashboard/buggys/buggys.component';
 import { AppRoutingModule } from './app-routing.module';
 import { AlternateGpsComponent } from './components/dashboard/alternate-gps/alternate-gps.component';
+import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
 
 import {GaugesModule} from 'ng-canvas-gauges';
 import { SpeedometerComponent } from './components/dashboard/speedometer/speedometer.component';
@@ -42,7 +43,8 @@ import { SpeedometerComponent } from './components/dashboard/speedometer/speedom
     BrowserModule,
     HttpClientModule,
     AppRoutingModule,
-    GaugesModule
+    GaugesModule,
+    NgbModule
   ],
   providers: [DataService],
   bootstrap: [AppComponent]
