@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { BuggysComponent } from './buggys.component';
 
@@ -6,7 +6,7 @@ describe('BuggysComponent', () => {
   let component: BuggysComponent;
   let fixture: ComponentFixture<BuggysComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ BuggysComponent ]
     })
