@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { SpeedometerComponent } from './speedometer.component';
 
@@ -6,7 +6,7 @@ describe('SpeedometerComponent', () => {
   let component: SpeedometerComponent;
   let fixture: ComponentFixture<SpeedometerComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ SpeedometerComponent ]
     })
