@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { BatteryComponent } from './battery.component';
 
@@ -6,7 +6,7 @@ describe('BatteryComponent', () => {
   let component: BatteryComponent;
   let fixture: ComponentFixture<BatteryComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ BatteryComponent ]
     })
